Enforce architecture boundaries with go-arch-lint
- Move goldmark usage from cli to markdown package
- Move migration detection from metamodel to repository
- Decouple markdown from graph (LoadSyncData returns data, repo populates graph)
- Add go-arch-lint config, CI job, and justfile target
