+++ internal/markdown/sync_test.go
@@ -6,7 +6,6 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/Sourcehaven-BV/rela/internal/graph"
 	"github.com/Sourcehaven-BV/rela/internal/metamodel"
 	"github.com/Sourcehaven-BV/rela/internal/model"
 	"github.com/Sourcehaven-BV/rela/internal/project"
@@ -14,8 +13,7 @@ import (
 	"github.com/Sourcehaven-BV/rela/internal/testutil"
 )
 
-func TestSyncFromFiles(t *testing.T) {
-	// Build test project with entities and relations
+func TestLoadSyncData(t *testing.T) {
 	entity1 := testutil.NewEntity("REQ-001", "requirement").
 		WithTitle("Requirement 1").
 		Build()
@@ -35,7 +33,7 @@ func TestSyncFromFiles(t *testing.T) {
 		WithRelation("addresses", "Addresses", []string{"decision"}, []string{"requirement"}).
 		Build()
 
-	testCtx, _, g := testutil.NewProject(t).
+	testCtx, _, _ := testutil.NewProject(t).
 		WithMetamodel(meta).
 		WithEntity(entity1).
 		WithEntity(entity2).
@@ -44,7 +42,6 @@ func TestSyncFromFiles(t *testing.T) {
 		WithRelation(rel2).
 		Build()
 
-	// Convert to project.Context
 	ctx := &project.Context{
 		Root:         testCtx.Root,
 		EntitiesDir:  testCtx.EntitiesDir,
@@ -52,169 +49,23 @@ func TestSyncFromFiles(t *testing.T) {
 		CacheDir:     testCtx.CacheDir,
 	}
 
-	// Sync
-	result, err := testIO.SyncFromFiles(ctx, meta, g)
+	data, err := testIO.LoadSyncData(ctx, meta)
 	testutil.AssertNoError(t, err)
 
-	// Verify results
-	testutil.AssertEqual(t, result.EntitiesLoaded, 3)
-	testutil.AssertEqual(t, result.RelationsLoaded, 2)
-	testutil.AssertLengthEqual(t, result.Errors, 0)
-
-	// Verify graph state
-	testutil.AssertEqual(t, g.NodeCount(), 3)
-	testutil.AssertEqual(t, g.EdgeCount(), 2)
-
-	// Verify specific entities exist
-	if _, ok := g.GetNode("REQ-001"); !ok {
-		t.Error("REQ-001 should exist in graph")
+	if len(data.Entities) != 3 {
+		t.Errorf("got %d entities, want 3", len(data.Entities))
 	}
-	if _, ok := g.GetNode("REQ-002"); !ok {
-		t.Error("REQ-002 should exist in graph")
+	if len(data.Relations) != 2 {
+		t.Errorf("got %d relations, want 2", len(data.Relations))
 	}
-	if _, ok := g.GetNode("DEC-001"); !ok {
-		t.Error("DEC-001 should exist in graph")
+	if len(data.Conflicted) != 0 {
+		t.Errorf("got %d conflicted, want 0", len(data.Conflicted))
 	}
 }
 
-func TestSyncFromFiles_MissingSourceEntity(t *testing.T) {
-	tmpDir := t.TempDir()
-	ctx := &project.Context{
-		Root:         tmpDir,
-		EntitiesDir:  filepath.Join(tmpDir, "entities"),
-		RelationsDir: filepath.Join(tmpDir, "relations"),
-	}
+func TestLoadSyncData_EmptyDirectories(t *testing.T) {
+	testCtx, meta, _ := testutil.NewProject(t).Build()
 
-	// Create only target entity
-	reqDir := filepath.Join(ctx.EntitiesDir, "requirement")
-	if err := os.MkdirAll(reqDir, 0755); err != nil {
-		t.Fatalf("failed to create dir: %v", err)
-	}
-	if err := os.MkdirAll(ctx.RelationsDir, 0755); err != nil {
-		t.Fatalf("failed to create relations dir: %v", err)
-	}
-
-	entityContent := `---
-id: REQ-001
-type: requirement
----
-`
-	if err := os.WriteFile(filepath.Join(reqDir, "REQ-001.md"), []byte(entityContent), 0644); err != nil {
-		t.Fatalf("failed to create entity: %v", err)
-	}
-
-	// Create relation with missing source
-	relationContent := `---
-from: DEC-999
-relation: addresses
-to: REQ-001
----
-`
-	relationPath := filepath.Join(ctx.RelationsDir, "DEC-999--addresses--REQ-001.md")
-	if err := os.WriteFile(relationPath, []byte(relationContent), 0644); err != nil {
-		t.Fatalf("failed to create relation: %v", err)
-	}
-
-	meta := &metamodel.Metamodel{
-		Entities: map[string]metamodel.EntityDef{
-			"requirement": {Label: "Requirement"},
-		},
-	}
-	g := graph.New()
-
-	result, err := testIO.SyncFromFiles(ctx, meta, g)
-	if err != nil {
-		t.Fatalf("SyncFromFiles failed: %v", err)
-	}
-
-	// Should have 1 entity, 0 relations, 1 error
-	if result.EntitiesLoaded != 1 {
-		t.Errorf("EntitiesLoaded = %d, want 1", result.EntitiesLoaded)
-	}
-	if result.RelationsLoaded != 0 {
-		t.Errorf("RelationsLoaded = %d, want 0 (missing source)", result.RelationsLoaded)
-	}
-	if len(result.Errors) != 1 {
-		t.Fatalf("got %d errors, want 1", len(result.Errors))
-	}
-
-	// Verify error message
-	if result.Errors[0].Error() != relationPath+": source entity not found: DEC-999" {
-		t.Errorf("error = %q, want source not found error", result.Errors[0].Error())
-	}
-}
-
-func TestSyncFromFiles_MissingTargetEntity(t *testing.T) {
-	tmpDir := t.TempDir()
-	ctx := &project.Context{
-		Root:         tmpDir,
-		EntitiesDir:  filepath.Join(tmpDir, "entities"),
-		RelationsDir: filepath.Join(tmpDir, "relations"),
-	}
-
-	// Create only source entity
-	decDir := filepath.Join(ctx.EntitiesDir, "decision")
-	if err := os.MkdirAll(decDir, 0755); err != nil {
-		t.Fatalf("failed to create dir: %v", err)
-	}
-	if err := os.MkdirAll(ctx.RelationsDir, 0755); err != nil {
-		t.Fatalf("failed to create relations dir: %v", err)
-	}
-
-	entityContent := `---
-id: DEC-001
-type: decision
----
-`
-	if err := os.WriteFile(filepath.Join(decDir, "DEC-001.md"), []byte(entityContent), 0644); err != nil {
-		t.Fatalf("failed to create entity: %v", err)
-	}
-
-	// Create relation with missing target
-	relationContent := `---
-from: DEC-001
-relation: addresses
-to: REQ-999
----
-`
-	relationPath := filepath.Join(ctx.RelationsDir, "DEC-001--addresses--REQ-999.md")
-	if err := os.WriteFile(relationPath, []byte(relationContent), 0644); err != nil {
-		t.Fatalf("failed to create relation: %v", err)
-	}
-
-	meta := &metamodel.Metamodel{
-		Entities: map[string]metamodel.EntityDef{
-			"decision": {Label: "Decision"},
-		},
-	}
-	g := graph.New()
-
-	result, err := testIO.SyncFromFiles(ctx, meta, g)
-	if err != nil {
-		t.Fatalf("SyncFromFiles failed: %v", err)
-	}
-
-	// Should have 1 entity, 0 relations, 1 error
-	if result.EntitiesLoaded != 1 {
-		t.Errorf("EntitiesLoaded = %d, want 1", result.EntitiesLoaded)
-	}
-	if result.RelationsLoaded != 0 {
-		t.Errorf("RelationsLoaded = %d, want 0 (missing target)", result.RelationsLoaded)
-	}
-	if len(result.Errors) != 1 {
-		t.Fatalf("got %d errors, want 1", len(result.Errors))
-	}
-
-	// Verify error message
-	if result.Errors[0].Error() != relationPath+": target entity not found: REQ-999" {
-		t.Errorf("error = %q, want target not found error", result.Errors[0].Error())
-	}
-}
-
-func TestSyncFromFiles_EmptyDirectories(t *testing.T) {
-	testCtx, meta, g := testutil.NewProject(t).Build()
-
-	// Convert to project.Context
 	ctx := &project.Context{
 		Root:         testCtx.Root,
 		EntitiesDir:  testCtx.EntitiesDir,
@@ -222,99 +73,21 @@ func TestSyncFromFiles_EmptyDirectories(t *testing.T) {
 		CacheDir:     testCtx.CacheDir,
 	}
 
-	result, err := testIO.SyncFromFiles(ctx, meta, g)
+	data, err := testIO.LoadSyncData(ctx, meta)
 	testutil.AssertNoError(t, err)
 
-	testutil.AssertEqual(t, result.EntitiesLoaded, 0)
-	testutil.AssertEqual(t, result.RelationsLoaded, 0)
-	testutil.AssertLengthEqual(t, result.Errors, 0)
-}
-
-func TestSyncFromFiles_ClearsExistingGraph(t *testing.T) {
-	tmpDir := t.TempDir()
-	ctx := &project.Context{
-		Root:         tmpDir,
-		EntitiesDir:  filepath.Join(tmpDir, "entities"),
-		RelationsDir: filepath.Join(tmpDir, "relations"),
+	if len(data.Entities) != 0 {
+		t.Errorf("got %d entities, want 0", len(data.Entities))
 	}
-
-	// Create entity directory
-	reqDir := filepath.Join(ctx.EntitiesDir, "requirement")
-	if err := os.MkdirAll(reqDir, 0755); err != nil {
-		t.Fatalf("failed to create dir: %v", err)
+	if len(data.Relations) != 0 {
+		t.Errorf("got %d relations, want 0", len(data.Relations))
 	}
-	if err := os.MkdirAll(ctx.RelationsDir, 0755); err != nil {
-		t.Fatalf("failed to create relations dir: %v", err)
-	}
-
-	// Create one entity
-	entityContent := `---
-id: REQ-001
-type: requirement
----
-`
-	if err := os.WriteFile(filepath.Join(reqDir, "REQ-001.md"), []byte(entityContent), 0644); err != nil {
-		t.Fatalf("failed to create entity: %v", err)
-	}
-
-	meta := &metamodel.Metamodel{
-		Entities: map[string]metamodel.EntityDef{
-			"requirement": {Label: "Requirement"},
-		},
-	}
-	g := graph.New()
-
-	// Pre-populate graph with different data
-	oldEntity := &model.Entity{
-		ID:         "OLD-001",
-		Type:       "old",
-		Properties: make(map[string]interface{}),
-	}
-	g.AddNode(oldEntity)
-
-	if g.NodeCount() != 1 {
-		t.Fatalf("pre-sync: graph has %d nodes, want 1", g.NodeCount())
-	}
-
-	// Sync should clear old data
-	result, err := testIO.SyncFromFiles(ctx, meta, g)
-	if err != nil {
-		t.Fatalf("SyncFromFiles failed: %v", err)
-	}
-
-	if result.EntitiesLoaded != 1 {
-		t.Errorf("EntitiesLoaded = %d, want 1", result.EntitiesLoaded)
-	}
-
-	// Old entity should be gone
-	if _, ok := g.GetNode("OLD-001"); ok {
-		t.Error("OLD-001 should not exist after sync (graph should be cleared)")
-	}
-
-	// New entity should exist
-	if _, ok := g.GetNode("REQ-001"); !ok {
-		t.Error("REQ-001 should exist after sync")
-	}
-
-	if g.NodeCount() != 1 {
-		t.Errorf("graph has %d nodes, want 1", g.NodeCount())
+	if len(data.Conflicted) != 0 {
+		t.Errorf("got %d conflicted, want 0", len(data.Conflicted))
 	}
 }
 
-func TestSyncError_Error(t *testing.T) {
-	err := &model.SyncError{
-		File:    "/path/to/file.md",
-		Message: "something went wrong",
-	}
-
-	expected := "/path/to/file.md: something went wrong"
-	if got := err.Error(); got != expected {
-		t.Errorf("Error() = %q, want %q", got, expected)
-	}
-}
-
-func TestSyncFromFiles_LoadEntitiesError(t *testing.T) {
-	// Create an ErrorFS that returns an error on Walk
+func TestLoadSyncData_LoadEntitiesError(t *testing.T) {
 	memFS := storage.NewMemFS()
 	errFS := storage.NewErrorFS(memFS)
 	errFS.WalkError = errors.New("permission denied")
@@ -332,18 +105,17 @@ func TestSyncFromFiles_LoadEntitiesError(t *testing.T) {
 			"requirement": {Label: "Requirement"},
 		},
 	}
-	g := graph.New()
 
-	_, err := errorIO.SyncFromFiles(ctx, meta, g)
+	_, err := errorIO.LoadSyncData(ctx, meta)
 	if err == nil {
-		t.Error("expected error from SyncFromFiles when LoadAllEntities fails")
+		t.Error("expected error from LoadSyncData when LoadAllEntities fails")
 	}
 	if err.Error() != "permission denied" {
 		t.Errorf("error = %q, want 'permission denied'", err.Error())
 	}
 }
 
-func TestSyncFromFiles_ConflictedFiles(t *testing.T) {
+func TestLoadSyncData_ConflictedFiles(t *testing.T) {
 	tmpDir := t.TempDir()
 	ctx := &project.Context{
 		Root:         tmpDir,
@@ -351,7 +123,6 @@ func TestSyncFromFiles_ConflictedFiles(t *testing.T) {
 		RelationsDir: filepath.Join(tmpDir, "relations"),
 	}
 
-	// Create entity directory
 	reqDir := filepath.Join(ctx.EntitiesDir, "requirement")
 	if err := os.MkdirAll(reqDir, 0755); err != nil {
 		t.Fatalf("failed to create dir: %v", err)
@@ -360,7 +131,7 @@ func TestSyncFromFiles_ConflictedFiles(t *testing.T) {
 		t.Fatalf("failed to create relations dir: %v", err)
 	}
 
-	// Create a valid entity
+	// Valid entity
 	validContent := `---
 id: REQ-001
 type: requirement
@@ -371,7 +142,7 @@ title: Valid Entity
 		t.Fatalf("failed to create valid entity: %v", err)
 	}
 
-	// Create a conflicted entity
+	// Conflicted entity
 	conflictedContent := `---
 id: REQ-002
 type: requirement
@@ -387,7 +158,7 @@ status: approved
 		t.Fatalf("failed to create conflicted entity: %v", err)
 	}
 
-	// Create a valid relation
+	// Valid relation
 	relationContent := `---
 from: REQ-001
 relation: depends-on
@@ -398,7 +169,7 @@ to: REQ-001
 		t.Fatalf("failed to create relation: %v", err)
 	}
 
-	// Create a conflicted relation
+	// Conflicted relation
 	conflictedRelationContent := `---
 from: REQ-001
 relation: addresses
@@ -419,31 +190,29 @@ to: REQ-003
 			"requirement": {Label: "Requirement"},
 		},
 	}
-	g := graph.New()
 
-	result, err := testIO.SyncFromFiles(ctx, meta, g)
+	data, err := testIO.LoadSyncData(ctx, meta)
 	if err != nil {
-		t.Fatalf("SyncFromFiles failed: %v", err)
+		t.Fatalf("LoadSyncData failed: %v", err)
 	}
 
 	// Should load 1 valid entity
-	if result.EntitiesLoaded != 1 {
-		t.Errorf("EntitiesLoaded = %d, want 1", result.EntitiesLoaded)
+	if len(data.Entities) != 1 {
+		t.Errorf("got %d entities, want 1", len(data.Entities))
 	}
 
 	// Should load 1 valid relation
-	if result.RelationsLoaded != 1 {
-		t.Errorf("RelationsLoaded = %d, want 1", result.RelationsLoaded)
+	if len(data.Relations) != 1 {
+		t.Errorf("got %d relations, want 1", len(data.Relations))
 	}
 
 	// Should track 2 conflicted files (1 entity + 1 relation)
-	if len(result.Conflicted) != 2 {
-		t.Errorf("got %d conflicted files, want 2", len(result.Conflicted))
+	if len(data.Conflicted) != 2 {
+		t.Errorf("got %d conflicted files, want 2", len(data.Conflicted))
 	}
 
-	// Verify the conflicted files are tracked
 	conflictedMap := make(map[string]bool)
-	for _, path := range result.Conflicted {
+	for _, path := range data.Conflicted {
 		conflictedMap[path] = true
 	}
 	if !conflictedMap[conflictedEntityPath] {
@@ -452,15 +221,16 @@ to: REQ-003
 	if !conflictedMap[conflictedRelationPath] {
 		t.Errorf("missing conflicted relation: %s", conflictedRelationPath)
 	}
+}
 
-	// Verify the graph only contains valid entities
-	if g.NodeCount() != 1 {
-		t.Errorf("graph has %d nodes, want 1", g.NodeCount())
+func TestSyncError_Error(t *testing.T) {
+	err := &model.SyncError{
+		File:    "/path/to/file.md",
+		Message: "something went wrong",
 	}
-	if _, ok := g.GetNode("REQ-001"); !ok {
-		t.Error("REQ-001 should exist in graph")
-	}
-	if _, ok := g.GetNode("REQ-002"); ok {
-		t.Error("REQ-002 should not exist in graph (conflicted)")
+
+	expected := "/path/to/file.md: something went wrong"
+	if got := err.Error(); got != expected {
+		t.Errorf("Error() = %q, want %q", got, expected)
 	}
 }
