+++ tickets/entities/features/FEAT-022.md
@@ -1,0 +1,8 @@
+---
+description: Enforce architecture boundaries between Go packages using go-arch-lint, preventing unauthorized cross-package imports
+id: FEAT-022
+status: implemented
+summary: Architecture boundary linting via go-arch-lint in CI
+title: Architecture boundary enforcement
+type: feature
+---
