+++ tickets/entities/tickets/TKT-019.md
@@ -1,0 +1,10 @@
+---
+description: Tighten architecture boundaries by decoupling markdown from graph, moving migration detection to repository, and removing goldmark from cli
+effort: s
+id: TKT-019
+kind: refactor
+priority: medium
+status: done
+title: Tighten architecture boundaries
+type: ticket
+---
