+++ tickets/relations/FEAT-022--requires--ci-pipeline.md
@@ -1,0 +1,5 @@
+---
+from: FEAT-022
+relation: requires
+to: ci-pipeline
+---
