+++ tickets/relations/TKT-019--affects--ci-pipeline.md
@@ -1,0 +1,5 @@
+---
+from: TKT-019
+relation: affects
+to: ci-pipeline
+---
